Converted getRows and dependent functions to async
Using callbacks turned out to be a nightmare with Views, so we went back to async functions.  getRows(), getFilteredRows(), getAllValues(), getNumericSpec() in GalyleoTable are now all async, as is Views.getData().  But the code is much simpler, and the clients still have an option with callback with the usual .then() syntax